requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
Werkzeug==3.0.1

# OCR Libraries (choose based on your preference)
# Tesseract OCR
pytesseract==0.3.10
Pillow==10.1.0

# PDF Processing
PyPDF2==3.0.1
pdf2image==1.16.3

# Image Processing
opencv-python==4.8.1.78
numpy==1.26.2

# Utilities
python-dotenv==1.0.0

# Multilingual Support
textblob==0.19.0
langdetect==1.0.9
reportlab==4.4.7
requests==2.32.5

# Testing
pytest==7.4.3
hypothesis==6.92.1
psutil==5.9.8
